Implement ThreeDayWeatherParser class

+++ app/src/main/java/com/ablethon/woongsang/gestcapturex/Parser/TodaysWeatherParser.java
@@ -39,7 +39,7 @@ public class TodaysWeatherParser extends DownloadTask {
             }
             avg = sum / obList.length();
 
-            String msg = "평균 기온: " + avg + "도, 최고 기온: " + max + "도, 최저 기온: " + min;
+            String msg = "평균 기온: " + (int)avg + "도, 최고 기온: " + (int)max + "도, 최저 기온: " + (int)min;
             WeatherActivity.myTTS.speak(msg, TextToSpeech.QUEUE_FLUSH, null);
 
             Log.i ("avg", Double.toString(avg));
